client: merge two api-related files
